Extract enums for products
