+++ tests/api/test_attributes.py
@@ -4,8 +4,8 @@ from django.db.models import Q
 from django.template.defaultfilters import slugify
 from tests.api.utils import get_graphql_content
 
-from saleor.graphql.product.types import (
-    AttributeTypeEnum, AttributeValueType, resolve_attribute_value_type)
+from saleor.graphql.product.enums import AttributeTypeEnum, AttributeValueType
+from saleor.graphql.product.types import resolve_attribute_value_type
 from saleor.graphql.product.utils import attributes_to_hstore
 from saleor.product.models import Attribute, AttributeValue, Category
